Add results CSV and all-pairs plot for Nexus 6
This is a set of results for all experiments from
https://github.com/aaaaalbert/benchmark_repy_v2 , run on a Google Nexus 6
with Android 5.1.1 via the `run-benchmark` script in this repo.

To save you some time, I'm also adding an all-pairs plot of the results.
The all-pairs plot shows the `log10` of the number of layers, time until
the loop over the wrapped call was started, elapsed time for the loop,
and elapsed time over number of layers (``relative''). The colors encode
the experiment spec, and the symbols the experiment category:
* black/red: "nocheck"/"handop", i.e. plain Python recursive calls,
* green: "seclayer", uses security layers for wrapping;
* naughts: "noarg" tests
* triangles: "immutarg" tests
* plusses: "mutarg" tests
* X'es: "noarg" tests

From the leftmost column of pairs plots:

It is seen that security layers take most time to set up, and the starttime
scales with the number of layers as well (e.g., 100 layers, at "2.0" in
the logscale graph, take around 100 second (simile) to set up); the pure
call-based recursion tests are O(1).

For the time elapsed, all tests scale with the number of layers, with
the seclayer ones performing slowest overall. The time elapsed per layer
is roughly constant, with sec layers about an order of magnitude slower
than the nocheck/handop tests. The seclayer and nocheck tests each
feature an approximately bimodal split. For nocheck, the noarg experiment
category consistently performs fastest; for seclayer, the category
does not explain the split fully (from visual inspection; your F may vary).