Add gyro-to-wav convert script
Also, add a readme, and some results from a Sony Tablet S and a Google
Nexus 6. As far as I can tell, there is nothing but noise in these
recordings, save for the peaks corresponding to the hand-clapping I
did at the beginning and end of the runs (in attempt to provide a
readable reference).